src/modules/module42/requirements.txt
streamlit
pymongo
pyjwt
bcrypt
streamlit-cookies-controller
faker
python-dotenv
fpdf2
